fix: resolves comments
