+++ app/lib/app/(public)/auth/register/register_page.dart
@@ -1,5 +1,5 @@
-import 'package:coopartilhar/app/(public)/auth/register/non_editable_register_item.dart';
-import 'package:coopartilhar/app/(public)/auth/register/register_item_field.dart';
+import 'package:coopartilhar/app/(public)/auth/register/widgets/non_editable_register_item.dart';
+import 'package:coopartilhar/app/(public)/auth/register/widgets/register_item_field.dart';
 import 'package:coopartilhar/app/features/register/interactor/controllers/register_controller.dart';
 import 'package:coopartilhar/injector.dart';
 import 'package:coopartilhar/routes.dart';
@@ -52,7 +52,7 @@ class _RegisterPageState extends State<RegisterPage> {
     final colors = CoopartilharColors.of(context);
     final textTheme = Theme.of(context).textTheme;
     final Size size = MediaQuery.sizeOf(context);
-    final keybordHeight = MediaQuery.of(context).viewInsets.bottom;
+    final keyboardHeight = MediaQuery.of(context).viewInsets.bottom;
 
     return Scaffold(
       appBar: AppBar(
@@ -62,92 +62,103 @@ class _RegisterPageState extends State<RegisterPage> {
           style: textTheme.displayLarge?.copyWith(color: colors.textColor),
         ),
         leading: IconButton(
-          icon: const Icon(Icons.chevron_left),
+          icon: Icon(UIcons.regularStraight.angle_small_left),
           onPressed: Navigator.of(context).pop,
         ),
         surfaceTintColor: Colors.transparent,
       ),
       resizeToAvoidBottomInset: false,
-      body: Container(
-        decoration: const BoxDecoration(
-          image: DecorationImage(
-            image: CooImages.cooBackgroundDetails,
-            fit: BoxFit.contain,
-            alignment: Alignment.bottomCenter,
+      body: LayoutBuilder(builder: (context, constraints) {
+        return Container(
+          decoration: const BoxDecoration(
+            image: DecorationImage(
+              image: CooImages.cooBackgroundDetails,
+              fit: BoxFit.contain,
+              alignment: Alignment.bottomCenter,
+            ),
           ),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24),
-          child: ValueListenableBuilder(
-            valueListenable: controller,
-            builder: (context, value, child) {
-              if (controller.value is LoadingState) {
-                return const Center(child: CircularProgressIndicator());
-              }
-              return SingleChildScrollView(
-                padding: EdgeInsets.only(bottom: keybordHeight),
-                child: Form(
-                  key: controller.formKey,
-                  child: SizedBox(
-                    height: size.height - kToolbarHeight,
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const SizedBox(height: 30),
-                        NonEditableRegisterItem(
-                          title: 'Nome completo',
-                          value: widget.name,
-                        ),
-                        const SizedBox(height: 30),
-                        NonEditableRegisterItem(
-                          title: 'CPF/CNPJ',
-                          value: widget.document,
-                        ),
-                        const SizedBox(height: 30),
-                        NonEditableRegisterItem(
-                          title: 'E-mail',
-                          value: widget.email,
-                        ),
-                        const SizedBox(height: 30),
-                        NonEditableRegisterItem(
-                          title: 'Telefone',
-                          value: widget.phone,
-                        ),
-                        const SizedBox(height: 30),
-                        RegisterItemField(
-                          title: 'Senha',
-                          controller: controller.passwordController,
-                          hint: 'Insira sua senha',
-                          isPassword: true,
-                          validator: controller.passwordValidator,
-                        ),
-                        const SizedBox(height: 30),
-                        RegisterItemField(
-                          title: 'Confirmar senha',
-                          controller: controller.repeatPasswordController,
-                          hint: 'Insira novamente sua senha',
-                          isPassword: true,
-                          validator: controller.repeatPasswordValidator,
-                        ),
-                        const Spacer(),
-                        CooButton.primary(
-                          label: 'Entrar',
-                          onPressed: () => controller.register(
-                            document: widget.document,
-                            name: widget.name,
-                            email: widget.email,
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 24),
+            child: ValueListenableBuilder(
+              valueListenable: controller,
+              builder: (context, value, child) {
+                if (controller.value is LoadingState) {
+                  return const Center(child: CircularProgressIndicator());
+                }
+                return Column(
+                  children: [
+                    Expanded(
+                      child: SingleChildScrollView(
+                        child: SizedBox(
+                          height: constraints.maxHeight,
+                          child: Form(
+                            key: controller.formKey,
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              mainAxisAlignment: keyboardHeight > 0 ? MainAxisAlignment.start : MainAxisAlignment.spaceBetween,
+                              children: [
+                                const SizedBox(height: 30),
+                                NonEditableRegisterItem(
+                                  title: 'Nome completo',
+                                  value: widget.name,
+                                ),
+                                const SizedBox(height: 30),
+                                NonEditableRegisterItem(
+                                  title: 'CPF/CNPJ',
+                                  value: widget.document,
+                                ),
+                                const SizedBox(height: 30),
+                                NonEditableRegisterItem(
+                                  title: 'E-mail',
+                                  value: widget.email,
+                                ),
+                                const SizedBox(height: 30),
+                                NonEditableRegisterItem(
+                                  title: 'Telefone',
+                                  value: widget.phone,
+                                ),
+                                const SizedBox(height: 30),
+                                RegisterItemField(
+                                  title: 'Senha',
+                                  controller: controller.passwordController,
+                                  hint: 'Insira sua senha',
+                                  isPassword: true,
+                                  validator: controller.passwordValidator,
+                                ),
+                                const SizedBox(height: 30),
+                                RegisterItemField(
+                                  title: 'Confirmar senha',
+                                  controller: controller.repeatPasswordController,
+                                  hint: 'Insira novamente sua senha',
+                                  isPassword: true,
+                                  validator: controller.repeatPasswordValidator,
+                                ),
+                                const Spacer(),
+                                CooButton.primary(
+                                  label: 'Entrar',
+                                  onPressed: () => controller.register(
+                                    document: widget.document,
+                                    name: widget.name,
+                                    email: widget.email,
+                                  ),
+                                ),
+                                const SizedBox(height: 40),
+                              ],
+                            ),
                           ),
                         ),
-                        const SizedBox(height: 40),
-                      ],
+                      ),
                     ),
-                  ),
-                ),
-              );
-            },
+                    SizedBox(
+                      height: keyboardHeight,
+                    )
+                  ],
+                );
+              },
+            ),
           ),
-        ),
-      ),
+        );
+      }),
     );
   }
 }
